Refine color matching process

+++ CompanyA/src/Core/Entities/ReferenceColorMatch.cs
@@ -2,7 +2,7 @@
 {
     public class ReferenceColorMatch
     {
-        public ReferenceColorMatch(ReferenceColor referenceColor, int confidence)
+        public ReferenceColorMatch(ReferenceColor referenceColor, double confidence)
         {
             ReferenceColor = referenceColor;
             Confidence = confidence;
@@ -10,6 +10,6 @@
 
         public ReferenceColor ReferenceColor { get; }
 
-        public int Confidence { get; }
+        public double Confidence { get; }
     }
 }
